# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real estate AI calling agent dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LeadInfo {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  propertyType?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bedrooms?: string | null;
}

export interface SentimentInfo {
  /** @nullable */
  score?: number | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  intent?: string | null;
}

export interface CallSummary {
  id: string;
  status: string;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  endedReason?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  summary?: string | null;
  sentiment?: SentimentInfo;
  lead?: LeadInfo;
  hasRecording: boolean;
}

export interface CallMessage {
  role: string;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  time?: number | null;
  /** @nullable */
  endTime?: number | null;
  /** @nullable */
  duration?: number | null;
}

export interface CallDetail {
  id: string;
  status: string;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  endedReason?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  recordingUrl?: string | null;
  /** @nullable */
  stereoRecordingUrl?: string | null;
  sentiment?: SentimentInfo;
  lead?: LeadInfo;
  followUpActions: string[];
  messages: CallMessage[];
  /** @nullable */
  cost?: number | null;
}

export interface SentimentBreakdown {
  positive: number;
  neutral: number;
  negative: number;
}

export interface DailyCallCount {
  date: string;
  count: number;
}

export interface CallStats {
  totalCalls: number;
  answeredCalls: number;
  /** @nullable */
  avgDurationSeconds?: number | null;
  sentimentBreakdown: SentimentBreakdown;
  callsByDay: DailyCallCount[];
  topIntents: string[];
  callsThisWeek: number;
  callsLastWeek: number;
}

export interface ClientRecord {
  id: number;
  clientCode: string;
  name: string;
  assistantId: string;
  /** @nullable */
  niche?: string | null;
  /** @nullable */
  clerkUserId?: string | null;
  isLinked: boolean;
  createdAt: string;
}

export interface ClientInput {
  clientCode: string;
  name: string;
  assistantId: string;
  /** @nullable */
  niche?: string | null;
}

export interface ClientUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  assistantId?: string | null;
  /** @nullable */
  niche?: string | null;
}

export interface OnboardInput {
  clientCode: string;
}

export interface AuthMeResponse {
  role: string;
  clerkUserId: string;
  /** @nullable */
  email?: string | null;
  isFirstUser: boolean;
}

export type ListCallsParams = {
limit?: number;
/**
 * @nullable
 */
createdAtGt?: string | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
assistantId?: string | null;
};

export type GetCallStatsParams = {
/**
 * @nullable
 */
assistantId?: string | null;
};