# lib/api-zod/src/generated/types/clientRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real estate AI calling agent dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientRecord {
  id: number;
  clientCode: string;
  name: string;
  assistantId: string;
  /** @nullable */
  niche?: string | null;
  /** @nullable */
  clerkUserId?: string | null;
  isLinked: boolean;
  createdAt: string;
}