# lib/api-zod/src/generated/types/clientUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real estate AI calling agent dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  assistantId?: string | null;
  /** @nullable */
  niche?: string | null;
}