# lib/api-zod/src/generated/types/onboardInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real estate AI calling agent dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardInput {
  clientCode: string;
}